Condition module: rename sync-condition->compare-engine + split CompareOperator/MinimumTextLength from PlanTerms (internal, behavior-preserving)
